refactor(parser): put mangarock models in their own file
